Make main.rb driver-agnostic.
The BAPS specific code has been factored out of main.rb into a new "driver"
class, which is dynamically loaded in via the config.yml file.  main.rb is now
completely clueless as to which playout system it is working with.

config.yml.example has been updated to have better documentation and reflect
the new driver model.

There is an outstanding bug with authentication that will be resolved next
commit.