Handle BAPS player states properly.
BAPS interprets a PLAY while PAUSEd as an instruction to return the position
marker to 0 before resuming play (the correct command to unpause is PAUSE), so
some jiggling of the SetPlayerState command was necessary to make this work.

SetPlayerState now expects a from-state as well as a to-state.

Channel commands now operate on channel IDs, not model objects.

Pausing a stopped channel no longer works (it doesn't make any sense!).