The great commands refactoring of 2013.
The BAPS command objects are now deprecated, as they are being replaced with a
Requester object (to mirror the Responder object).

The Requester both contains methods for directly banging BAPS with login
information, and also deals with the assignment of Handlers to model objects
via configuring the model creator.

Model objects now take one handler object as opposed to multiple handler
functions.  Handlers are attached by consulting the model object's
handler_target and matching it against the target on the handler itself.

bra now loudly proclaims (on stdout, for now) whenever a model object is
without a handler, for debug/implementation purposes.

TL;DR handlers are better now.