Create the PayloadProcessor.
This factors out the handling of payloads for both PUT and POST, both as in
stripping off an index->item wrapper and as in deciding whether the payload is
a URL, a hash representation, to be sent for forwarding to another handler, or
invalid.

This is part of the forthcoming move of the handler architecture to a
driver-neutral zone.