Revamping the Inspector.
The API Inspector now has its own class, which contains all of the state that
the HAML needs in a more convenient manner than as previous.

Much of the API inspector HAML files have been rewritten in parts to make use
of the changed inspector framework.

Also, PlaylistSet and PlayerSet have been replaced with a single class, Set.